requirements.txt
kazoo==2.6.1
psycopg2-binary==2.8.4
lockfile==0.12.2
python-daemon==2.1.1
pyyaml==5.3.1
